Use L-BFGS-B Fortran library for native logistic regression benchmark (#8)
This PR swaps out the native logistic regression benchmark's solver for the same one used in SciPy. L-BFGS-B is wrapped in a minimal DAAL optimization_solver class and directly set as the solver for DAAL's logistic regression. This brings native performance for this benchmark to match our optimized scikit-learn performance for a sample binary classification problem with 100k samples and 1k features.

Because we must now directly call BLAS and LAPACK functions instead of using DAAL's internal BLAS and LAPACK implementations, we add a dependency on MKL as well.